Add basic settlement functionality
TODO: Resources need to be removed in the future when economy gets updated, UI still needs to be setup however basic functionality is that you can add buildings in code and can click on the settlement in the world. Resources still need to be collected from buildings per turn but most of the framework is there.